wip(PluginSystem): enabling aditional resources for plugins

+++ MiAZ/backend/pluginsystem.py
@@ -96,25 +96,32 @@ class MiAZPluginManager(GObject.GObject):
             azip.extractall(ENV['LPATH']['PLUGINS'])
             self.engine.rescan_plugins()
             config = self.app.get_config('Plugin')
-            config.add_available(key=fn1_name)
+            config.add_available(key=plugin_name)
             plugin_fname = os.path.basename(plugin_path)
-            self.log.debug(f"Plugin '{plugin_fname}' added to 'ENV['LPATH']['PLUGINS']'")
+            self.log.debug(f"Plugin '{plugin_fname}' added to '{ENV['LPATH']['PLUGINS']}'")
         # ~ self.emit('plugins-updated')
         return valid
 
     def remove_plugin(self, plugin: Peas.PluginInfo):
         """Remove plugin for user space plugins"""
-        # FIXME: Make sure the plugin is deleted and unloaded
-        ENV = self.app.get_env()
-        module = plugin.get_module_name()
-        self.unload_plugin(plugin)
-        plugin_head = os.path.join(ENV['LPATH']['PLUGINS'], f'{module}.plugin')
-        plugin_body = os.path.join(ENV['LPATH']['PLUGINS'], f'{module}.py')
-        os.unlink(plugin_head)
-        os.unlink(plugin_body)
         config = self.app.get_config('Plugin')
-        config.remove_available(key=module)
-        return True
+        module = plugin.get_module_name()
+        if not config.exists_used(module):
+            self.log.debug(f"Plugin '{module}' is not being used and will be deleted")
+            utils = self.app.get_service('util')
+            ENV = self.app.get_env()
+            self.unload_plugin(plugin)
+            plugin_head = os.path.join(ENV['LPATH']['PLUGINS'], f'{module}.plugin')
+            plugin_body = os.path.join(ENV['LPATH']['PLUGINS'], f'{module}.py')
+            os.unlink(plugin_head)
+            os.unlink(plugin_body)
+            plugin_res = os.path.join(ENV['LPATH']['PLUGRES'], module)
+            utils.directory_remove(plugin_res)
+            config.remove_available(key=module)
+            return True
+        else:
+            self.log.warning(f"Plugin {module} can't be deleted because it is still in use")
+            return False
 
     def rescan_plugins(self):
         try:
@@ -220,14 +227,10 @@ class MiAZPluginManager(GObject.GObject):
         # User plugins
         # All user space plugins are available for all repositories
         # However, each repository can use none, any or all of them
-
-        # FIXME: User plugins disabled temporary
-        # ~ if not os.path.exists(ENV['LPATH']['PLUGINS']):
-            # ~ os.makedirs(ENV['LPATH']['PLUGINS'], exist_ok=True)
-        # ~ self.engine.add_search_path(ENV['LPATH']['PLUGINS'])
-        # ~ self.log.debug(f"Added user plugins dir: {ENV['LPATH']['PLUGINS']}")
-
-
+        if not os.path.exists(ENV['LPATH']['PLUGINS']):
+            os.makedirs(ENV['LPATH']['PLUGINS'], exist_ok=True)
+        self.engine.add_search_path(ENV['LPATH']['PLUGINS'])
+        self.log.debug(f"Added user plugins dir: {ENV['LPATH']['PLUGINS']}")
 
     @staticmethod
     def __extension_removed_cb(unused_set, unused_plugin_info, extension):
